+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 1.3.0
 * Allow get status from `H.Response` #15
 * Minor performance improvement
+* Add `H.KV.entrySet()` method
 
 1.2.3 - 15/Jan/2018
 * Minor performance improvement